rough functionality for admin delete users
added rough functionality for an admin to delete a user, right now it only works on users without habits, and will spit the admin to a JSON dump after they delete a user. Currently working on a fix, will probably use JS scripting in the html to achieve this, the same way I did the dynamic habits list.