Added full CRUD functionality to Admin
Added routes for Edit User, Add User, and Fixed Delete User.

Decided to drop JSonify for this task, as it was complicating things. Plan to re-implement.

Edit is implemented with a bit of JS, which loads a small text area for the Admin to be able to use it. The Create is implemented almost the same way as login.