Added update and delete functionality to habits
Added two buttons to the right of each habit in the list: Edit and Delete
When Edit is clicked, it calls showEditPopup(), which makes an entry box visible. User can enter the new description for the habit, and either press ok or cancel.
Also created javascript methods closeEditPopup() and submitEdit() in UserDashboard.html, and route /edithabit in server.py to make this functionality possible
When Delete is clicked, the habit is deleted from the database (route /deletehabit) and the UI is updated (deleteHabit() javascript method)