UI updates + Plotly integration + HS updates
Adjusted styling in dashboard.css to make it more pleasing to the eye.

Adjusted multiple blocks in
UserDashboard.html to align blocks better, and make room for graphs in plotly.

Integrated a bar graph to track progess on your daily habits. Added a space to add the Macro line chart, in development.

Will probably add a GraphService.py, where graphs will be made fully in the backend, and passed into the ViewModel (Routes)